+++ database/migrations/2024_12_03_235408_create_contents_table.php
@@ -22,10 +22,6 @@ return new class extends Migration
             $table->foreignId('entity_id')->nullable(false)->constrained(CoreTables::Entities->value, 'id', "{$table_name}_entity_id_FK")->cascadeOnDelete()->comment('The entity that the content belongs to');
             $table->foreignId('presettable_id')->nullable(false)->constrained(CoreTables::Presettables->value, 'id', "{$table_name}_presettable_id_FK")->cascadeOnDelete()->comment('The entity preset that the content belongs to');
             $table->json('shared_components')->nullable()->comment('The shared dynamic components of the content');
-            $table->string('origin_label')->nullable()
-                ->comment('Human-readable name of the external origin source');
-            $table->string('origin_url', 2048)->nullable()
-                ->comment('Link to the external origin source');
             $table->integer('order_column')->nullable(false)->default(0)->index("{$table_name}_order_column_IDX")->comment('The order of the content');
 
             MigrateUtils::timestamps(